The routes for getting the data are defined in a separate file "get_data.py"
